requirements.txt
streamlit
cryptography
requests
clipboard
pynput
pillow
opencv-python
